Accelerated ES6 JS Training
The training is made by Maximilian Schwarzmüller, but I made it in React to be more simple to visualize.
I've arranged it based on chapters(2-11), every file in the chapter is imported in App.js (commented).
Where was needed, I added explanations in the comments.
Chapter 11 is the Weather Project.